Add CSS transition support to virtual-dom.
The class ".transition-entering" is added to a node when it is being
inserted into the DOM and will be removed once it has been inserted.
This enables us to trigger css transitions on node creation.

As an example I've created a slideshow app in Elm using these new
changes. This example can be found here:
https://github.com/arianvp/elm-gallery

I've implemented the mechanism as discussed here:
https://github.com/Matt-Esch/virtual-dom/issues/104#issuecomment-73732749

There have been some thoughts on how to do transitions on node deletion
but there hasn't been a solid answer yet. It involves listening to the
"transitionend" event on a node and setting the model to a consistent
state at that point.  Those are all changes that should be in the user's
code and not in this library.