Fixes #9 - Virtual DOM updates don't work when embedding Graphics.Element into Html
Changed the "fromElement" implementation to use shared "init" and "update" methods
via a prototype. This is because the virtual-dom library expects these methods to
be shared and compares the "init" methods of two embedded widgets to check whether
they're of the same type.

This implementation is likely to be more efficient than the earlier one since each
fromElement call now doesn't create two closures.