Start moving towards "event node" approach to routing events
This approach means that `map` will add an “event node” to a certain
node. An “event node” carries a tagger and a pointer to its parent
event node. So when there is a diff of a `map` node, the patch will
just be updating the relevant event node. If a `map` node is created or
destroyed, you will need a full re-render of the stuff below to correct
the parent references of events.