Remove old concept of keys
The existing API clashes with laziness optimizations, so this is moving
towards a new style.

If you have a big list, you’d likely want to make each entry lazy. It
would suck to have to build 1000 virtual nodes. If you want keys, you
need to add them as an attribute, but lazy takes no attributes! This
means you need to add an extra DOM node with a key where it has one
child that’s lazy.

The plan is for it to not be this way!