Switch to more compact representation of facts
When there are few facts, the object should be a decent bit smaller. As
the number of facts grows larger, it should be no worse than the old
way.

This also makes `style` a primitive of its own. This should be a little
bit faster (allocating one fewer list and skipping some function
calls). It also means if you have two different styles specified, you
will get all of the styles put together.